refactor: extract hotkeys composable and floating controls component
- Implemented Phase 5 of the modularization roadmap (ROADMAP.md).
- Created `useHotkeys` composable in `src/composables/useHotkeys.ts` to manage global keydown shortcuts.
- Created `FloatingControls` component in `src/components/FloatingControls.vue` to contain the floating action buttons (Switch Vocalist, Word End, Back, Next).
- Cleaned up `src/App.vue` by removing inline floating action markup and hotkey listeners.
